+++ example/lib/main.dart
@@ -124,8 +124,8 @@ class _MyAppState extends State<MyApp> {
                       },
                     ),
                     RaisedButton(
-                      child: const Text(
-                          'Send referral URL for this installation'),
+                      child:
+                          const Text('Send referral URL for this installation'),
                       onPressed: () {
                         /// Sets referral URL for this installation. This might
                         /// be required to track some specific traffic