+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.6
+
+* Upgrade AppMetrica SDK (Android 3.18.0/iOS 3.14.0).
+* Upgrade Android plugin.
+
 ## 0.0.5+2
 
 * Upgrade AppMetrica SDK (Android 3.14.3/iOS 3.11.1).